pyproject.toml
[project]
name = "fvapps"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Quinn Dougherty", email = "quinnd@riseup.net" },
    { name = "Ronak Mehta", email = "ronakrm@gmail.com" }
]
dependencies = [
    "datasets>=2.21.0",
    "anthropic>=0.34.2",
    "python-dotenv>=1.0.1",
    "openai>=1.44.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
hypothesis_gen_example = "scripts.example_agent_run:python_main"
lean_gen_example = "scripts.example_agent_run:lean_main"
apps_preprocess_test = "scripts.apps_preproc:test"
apps_preprocess_all = "scripts.apps_preproc:main_all"

[tool.rye]
managed = true
dev-dependencies = [
    "jupyter>=1.1.1",
    "black>=24.8.0",
    "pytest>=8.3.2",
    "hypothesis>=6.112.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/benchmark", "src/baselines", "src/scripts"]

[tool.pytest.ini_options]
testpaths = [
  "test"
]
